pyproject.toml
[project]
name = "hans-ivp-ivr"
version = "0.1.0"
description = ""
authors = [
    {name = "Danylo",email = "dchynonyi@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas (>=2.3.3,<3.0.0)",
    "vectorbt (>=0.28.2,<0.29.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "matplotlib (>=3.10.8,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
